ENH add y6 piff class
