+++ eastlake/__init__.py
@@ -11,3 +11,4 @@ except PackageNotFoundError:
 
 from .pipeline import register_pipeline_step  # noqa
 from . import des_piff  # noqa
+from . import des_smoothpiff  # noqa
